[#30]refactor: 코드 구조 수정
