+++ src/app/gathering/detail/[id]/page.tsx
@@ -4,6 +4,8 @@ import {
   QueryClient,
 } from "@tanstack/react-query";
 
+import FetchBoundary from "@/components/boundary/FetchBoundary";
+import GatheringReviewSkeleton from "@/components/skeletons/gathering-detail/GatheringReviewSkeleton";
 import { gatheringDetailQueryOption } from "@/hooks/gathering/useGatheringDetail";
 import { participantsQueryOption } from "@/hooks/gathering/useParticipants";
 import { reviewsByGatheringIdQueryOption } from "@/hooks/review/useReviewsByGatheringId";
@@ -36,7 +38,9 @@ export default async function GatheringDetailPage({
       <div className="base-wrap bg-gray-50">
         <article className="flex h-full flex-col gap-8 px-0 pt-5 md:px-20 md:pt-10">
           <GatheringInformation id={id} />
-          <GatheringReviewBoard />
+          <FetchBoundary fallback={<GatheringReviewSkeleton />}>
+            <GatheringReviewBoard id={id} />
+          </FetchBoundary>
         </article>
         <FloatingBar id={id} />
       </div>
