+++ src/constants/review.ts
@@ -1,12 +1,8 @@
 import { v4 as uuidv4 } from "uuid";
 
-export interface IMainReview {
-  id: string;
-  content: string;
-  rating: number;
-}
+import { IMainReview } from "@/types/review/main-review";
 
-export const reviewList: IMainReview[] = [
+export const mainReviewList: IMainReview[] = [
   {
     id: uuidv4(),
     content: `많은 사람들과 함께 \n 좋아하는 것을 즐길 수 있어 \n 좋았어요.`,
@@ -15,7 +11,7 @@ export const reviewList: IMainReview[] = [
   {
     id: uuidv4(),
     content: `같은 취향을 가진 사람들과 \n 소통할 수 있어 즐거웠어요`,
-    rating: 4,
+    rating: 4.5,
   },
   {
     id: uuidv4(),
@@ -25,12 +21,12 @@ export const reviewList: IMainReview[] = [
   {
     id: uuidv4(),
     content: `함께 몰입할 수 있어 즐거웠어요`,
-    rating: 4,
+    rating: 4.5,
   },
   {
     id: uuidv4(),
     content: `다 같이 즐기니까 \n 새로운 매력을 발견했어요`,
-    rating: 4,
+    rating: 5,
   },
   {
     id: uuidv4(),
