+++ src/types/review/main-review.ts
@@ -1,0 +1,5 @@
+export interface IMainReview {
+  id: string;
+  content: string;
+  rating: number;
+}
